backend/search/requirements.txt
fastapi
uvicorn
python-dotenv
kdbai-client
voyageai
transformers
pandas
numpy
pydantic
unidecode
psycopg2-binary
